added contracts handler file
